more improvements. Adapted daily rewards plugin for checking in server side. Added adcolony V4VC videos.

+++ Assets/GameTemplate/Scripts/Common/Utils/UI/UINumberProgressEffect.cs
@@ -9,27 +9,27 @@ using System.Collections.Generic;
 [RequireComponent(typeof(Text))]
 [RequireComponent(typeof(Button))]
 public class UINumberProgressEffect : UIBaseButton {
-	public int currentNumber = 0;
-	public int targetNumber = 100;
+	public long currentNumber = 0;
+	public long targetNumber = 100;
 	public float percentagIncrement = 0.3f;
-
+	
 	[Tooltip("True if a window handle the typing initialization")]
 	public bool managedByWindow = true;
 	public bool initStoped = false;
 	public bool allowStop = true;
 	public float effectPause = 0.2f;
 	public List<string> soundIDs;
-//	public AudioClip[] sounds;
+	//	public AudioClip[] sounds;
 	public bool loopMessage = false;
 	public bool emptyWhenLoop = true;
-
+	
 	
 	private string message;
 	private string stoppedMessage;
 	private Text lbMessage;
 	private bool stopType = false;
 	private bool initedTyping = false;
-
+	
 	public bool isLooped{
 		get{return loopMessage;}
 	}
@@ -41,11 +41,11 @@ public class UINumberProgressEffect : UIBaseButton {
 	void Start(){
 		if(!initStoped && !managedByWindow)
 			initType(currentNumber, targetNumber);
-
+		
 		percentagIncrement = Mathf.Clamp(percentagIncrement, 0.1f, 1f);
 	}
-
-
+	
+	
 	
 	private void init(){
 		lbMessage = GetComponent<Text>();
@@ -54,22 +54,22 @@ public class UINumberProgressEffect : UIBaseButton {
 		lbMessage.text = "";
 		stopType = false;
 		
-//		if(initStoped){
-//			lbMessage.text = stoppedMessage;
-//		}
+		//		if(initStoped){
+		//			lbMessage.text = stoppedMessage;
+		//		}
 	}
 	
 	IEnumerator TypeText () {
 		if(emptyWhenLoop)
 			lbMessage.text = "";
-
-		int totalStepsByOne = (targetNumber-currentNumber); //by one step
-		int increment = (int) (totalStepsByOne*percentagIncrement);
-
+		
+		long totalStepsByOne = (targetNumber-currentNumber); //by one step
+		long increment = (long) (totalStepsByOne*percentagIncrement);
+		
 		if(increment > 0){
-			int totalSteps = totalStepsByOne/increment;
-
-
+			long totalSteps = totalStepsByOne/increment;
+			
+			
 			for (int i=0; i<totalSteps; i++) {
 				lbMessage.text = currentNumber.ToString();
 				
@@ -78,19 +78,19 @@ public class UINumberProgressEffect : UIBaseButton {
 					string id = soundIDs[soundIndex];
 					BaseSoundManager.Instance.play(id);
 				}
-	//			else if (sounds != null && sounds.Length > 0 && PlayerPrefs.GetFloat(GameSettings.PP_SOUND) > 0f){
-	//				GetComponent<AudioSource>().PlayOneShot(sounds[Random.Range(0, sounds.Length)]);
-	//			}
+				//			else if (sounds != null && sounds.Length > 0 && PlayerPrefs.GetFloat(GameSettings.PP_SOUND) > 0f){
+				//				GetComponent<AudioSource>().PlayOneShot(sounds[Random.Range(0, sounds.Length)]);
+				//			}
 				
 				yield return 0;
 				yield return StartCoroutine(TimeUtils.WaitForRealSeconds( effectPause ) );
-
+				
 				if(currentNumber<targetNumber)
 					currentNumber += increment;
-
+				
 				if(stopType)
 					break;
-
+				
 			}      
 			
 			if(loopMessage && !stopType){
@@ -107,17 +107,17 @@ public class UINumberProgressEffect : UIBaseButton {
 		}
 	}
 	
-	public virtual void initType(int initialValue, int nextValue){
+	public virtual void initType(long initialValue, long nextValue){
 		currentNumber = initialValue;
 		targetNumber = nextValue;
-
+		
 		init();
 		
 		stopType = false;
 		StartCoroutine(TypeText ());
 		initedTyping = true;
 	}
-
+	
 	public virtual void stop(){
 		if(!stopType){
 			stopType = true;
@@ -125,9 +125,9 @@ public class UINumberProgressEffect : UIBaseButton {
 			
 			lbMessage.text = stoppedMessage;
 			
-//			if (sounds != null && sounds.Length > 0 && PlayerPrefs.GetFloat(GameSettings.PP_SOUND) > 0f){
-//				GetComponent<AudioSource>().PlayOneShot(sounds[Random.Range(0, sounds.Length)]);
-//			}
+			//			if (sounds != null && sounds.Length > 0 && PlayerPrefs.GetFloat(GameSettings.PP_SOUND) > 0f){
+			//				GetComponent<AudioSource>().PlayOneShot(sounds[Random.Range(0, sounds.Length)]);
+			//			}
 		}
 	}
 	
@@ -140,6 +140,6 @@ public class UINumberProgressEffect : UIBaseButton {
 		base.doPress ();
 		stop();
 	}
-
-
+	
+	
 }
